ia/requirements.txt

python-dotenv
google-generativeai
google-api-python-client
